Refactor editorialPalette.ts into dir with index.ts
