+++ common-rendering/src/editorialPalette/background.ts
@@ -1,0 +1,139 @@
+// ----- Imports ----- //
+
+import {
+	ArticleFormat,
+	ArticleDisplay,
+	ArticleDesign,
+	ArticlePillar,
+} from '@guardian/libs';
+import {
+	neutral,
+	culture,
+	sport,
+	lifestyle,
+	opinion,
+	news,
+} from '@guardian/source-foundations';
+import { Colour } from '.';
+
+// ----- Functions ----- //
+
+const headline = (format: ArticleFormat): Colour => {
+	if (format.display === ArticleDisplay.Immersive) {
+		return neutral[7];
+	} else if (format.design === ArticleDesign.DeadBlog) {
+		return neutral[97];
+	} else if (format.design === ArticleDesign.LiveBlog) {
+		switch (format.theme) {
+			case ArticlePillar.Culture:
+				return culture[300];
+			case ArticlePillar.Sport:
+				return sport[300];
+			case ArticlePillar.Lifestyle:
+				return lifestyle[300];
+			case ArticlePillar.Opinion:
+				return opinion[300];
+			case ArticlePillar.News:
+				return news[300];
+			default:
+				return news[300];
+		}
+	} else if (
+		format.design === ArticleDesign.Comment ||
+		format.design === ArticleDesign.Letter ||
+		format.design === ArticleDesign.Editorial
+	) {
+		return opinion[800];
+	} else if (format.design === ArticleDesign.Media) {
+		return neutral[10];
+	}
+
+	return neutral[100];
+};
+
+const headlineDark = (format: ArticleFormat): Colour => {
+	if (
+		format.design === ArticleDesign.LiveBlog ||
+		format.design === ArticleDesign.DeadBlog
+	) {
+		switch (format.theme) {
+			case ArticlePillar.Culture:
+				return culture[200];
+			case ArticlePillar.Sport:
+				return sport[200];
+			case ArticlePillar.Lifestyle:
+				return lifestyle[200];
+			case ArticlePillar.Opinion:
+				return opinion[200];
+			case ArticlePillar.News:
+				return news[200];
+			default:
+				return news[200];
+		}
+	}
+	return neutral[10];
+};
+
+const standfirst = ({ design, theme }: ArticleFormat): Colour => {
+	if (design === ArticleDesign.DeadBlog) {
+		return neutral[93];
+	}
+
+	if (design === ArticleDesign.LiveBlog) {
+		switch (theme) {
+			case ArticlePillar.Opinion:
+				return opinion[200];
+			case ArticlePillar.Sport:
+				return sport[100];
+			case ArticlePillar.Culture:
+				return culture[200];
+			case ArticlePillar.Lifestyle:
+				return lifestyle[200];
+			case ArticlePillar.News:
+			default:
+				return news[200];
+		}
+	}
+
+	if (design === ArticleDesign.Comment) {
+		return opinion[800];
+	}
+
+	return neutral[100];
+};
+
+const standfirstDark = ({ design, theme }: ArticleFormat): Colour => {
+	switch (design) {
+		case ArticleDesign.DeadBlog:
+			return neutral[10];
+		case ArticleDesign.LiveBlog:
+			switch (theme) {
+				case ArticlePillar.Opinion:
+					return opinion[100];
+				case ArticlePillar.Sport:
+					return sport[100];
+				case ArticlePillar.Culture:
+					return culture[100];
+				case ArticlePillar.Lifestyle:
+					return lifestyle[100];
+				case ArticlePillar.News:
+				default:
+					return news[100];
+			}
+		default:
+			return neutral[10];
+	}
+};
+
+// ----- API ----- //
+
+const background = {
+	headline,
+	headlineDark,
+	standfirst,
+	standfirstDark,
+};
+
+// ----- Exports ----- //
+
+export { background };
